+++ Source/RadiantPi.Sony.Cledis/ASonyCledisClient.cs
@@ -53,10 +53,12 @@ namespace RadiantPi.Sony.Cledis {
         public abstract Task SetInputAsync(SonyCledisInput input);
         public abstract Task SetPictureModeAsync(SonyCledisPictureMode mode);
         public abstract Task SetPowerAsync(SonyCledisPower power);
-        public abstract Task Set2D3DSelectionAsync(SonyCledis2D3D selection);
+        public abstract Task Set2D3DModeAsync(SonyCledis2D3DMode mode);
         public abstract Task Set3DFormatAsync(SonyCledis3DFormat format);
-        public abstract Task SetDualDisplayPort3D4KAsync(SonyCledisDualDisplayPort3D4K status);
+        public abstract Task SetDualDisplayPort3D4KModeAsync(SonyCledisDualDisplayPort3D4KMode mode);
         public abstract Task SetFanModeAsync(SonyCledisFanMode mode);
+        public abstract Task SetHorizontalPictureShiftAsync(SonyCledisInput input, int shift);
+        public abstract Task SetVerticalPictureShiftAsync(SonyCledisInput input, int shift);
 
         //--- Methods ---
         protected T ConvertResponse<T>(string response) {