+++ ReleaseNotes.md
@@ -4,10 +4,12 @@
 
 ### Features
 
-* Added `Set2D3DSelectionAsync()` to select either 2D or 3D operation.
+* Added `Set2D3DModeAsync()` to select either 2D or 3D operation.
 * Added `Set3DFormatAsync()` to select 3D image format.
-* Added `SetDualDisplayPort3D4KAsync()` to enable/disable 3D 4K mode.
+* Added `SetDualDisplayPort3D4KModeAsync()` to enable/disable 3D 4K mode.
 * Added `SetFanModeAsync()` to control the fan speed.
+* Added `SetHorizontalPictureShiftAsync()` to set the horizontal shift of the image by input.
+* Added `SetVerticalPictureShiftAsync()` to set the vertical shift of the image by input.
 
 ## v1.0 (2021-12-07)